chore: remove old implementations
