+++ src/exports/e2e/managers/index.ts
@@ -1,14 +1,3 @@
 export { StateManager } from "./state-manager";
 export { ENV } from "./env-manager";
-export { 
-  createContractsManager,
-  createAppContract,
-  createAppViewContract,
-  createUserContract,
-  createUserViewContract
-} from "./contracts-manager";
-export { 
-  createChainClient,
-  ChainClient,
-  PublicViemClientManager
-} from "./chain-client";
+export { ChainClient } from "./chain-client";
