+++ src/exports/e2e/managers/state-manager.ts
@@ -2,9 +2,10 @@ import chalk from "chalk";
 import { promises as fs } from "fs";
 import path from "path";
 import { createHash } from "crypto";
-import { generatePrivateKey, privateKeyToAccount } from "viem/accounts";
+import { ethers } from "ethers";
 import { CapacityCreditInfo } from "../utils/mint-cc";
 import { PKPInfo } from "../utils/mint-pkp";
+import { PermissionData } from "@lit-protocol/vincent-contracts-sdk";
 
 interface AccountState {
   privateKey: string;
@@ -24,22 +25,20 @@ interface CapacityCreditState extends CapacityCreditInfo {
 }
 
 interface VincentAppState {
-  appId: string;
-  appVersion: string;
+  appId: number;
+  appVersion: number;
   createdAt: string;
   network: string;
   delegateeAddress: string;
   abilityIpfsCids: string[];
   abilityPolicies: string[][];
-  abilityPolicyParameterNames: string[][];
-  abilityPolicyParameterTypes: number[][];
-  abilityPolicyParameterValues?: string[][]; // Optional for backward compatibility
+  policyParams: PermissionData;
 }
 
 interface PKPAppPermissionState {
-  pkpTokenId: string;
-  appId: string;
-  appVersion: string;
+  pkpEthAddress: string;
+  appId: number;
+  appVersion: number;
   permittedAt: string;
   network: string;
 }
@@ -57,8 +56,8 @@ interface ConfigurationState {
 }
 
 interface AppVersionConfiguration {
-  appId: string;
-  appVersion: string;
+  appId: number;
+  appVersion: number;
   abilityIpfsCids: string[];
   abilityPolicyCids: string[][];
   lastUsed: string;
@@ -283,47 +282,47 @@ export class StateManager {
    * Generate app version key from appId and appVersion
    */
   private static generateAppVersionKey(
-    appId: string,
-    appVersion: string
+    appId: number,
+    appVersion: number
   ): string {
     return `${appId}-${appVersion}`;
   }
 
-  /**
-   * Get app version configuration, creating if needed
-   */
-  private getAppVersionConfiguration(
-    appId: string,
-    appVersion: string
-  ): AppVersionConfiguration {
-    const appVersionKey = StateManager.generateAppVersionKey(appId, appVersion);
+  // /**
+  //  * Get app version configuration, creating if needed
+  //  */
+  // private getAppVersionConfiguration(
+  //   appId: string,
+  //   appVersion: string
+  // ): AppVersionConfiguration {
+  //   const appVersionKey = StateManager.generateAppVersionKey(appId, appVersion);
 
-    if (!this.nestedState.testFiles[this.testFileName]) {
-      this.nestedState.testFiles[this.testFileName] = {
-        configurations: {},
-        appVersions: {},
-      };
-    }
+  //   if (!this.nestedState.testFiles[this.testFileName]) {
+  //     this.nestedState.testFiles[this.testFileName] = {
+  //       configurations: {},
+  //       appVersions: {},
+  //     };
+  //   }
 
-    if (
-      !this.nestedState.testFiles[this.testFileName].appVersions[appVersionKey]
-    ) {
-      throw new Error(
-        `App version configuration ${appVersionKey} not found. It should be created when saving the app.`
-      );
-    }
+  //   if (
+  //     !this.nestedState.testFiles[this.testFileName].appVersions[appVersionKey]
+  //   ) {
+  //     throw new Error(
+  //       `App version configuration ${appVersionKey} not found. It should be created when saving the app.`
+  //     );
+  //   }
 
-    return this.nestedState.testFiles[this.testFileName].appVersions[
-      appVersionKey
-    ];
-  }
+  //   return this.nestedState.testFiles[this.testFileName].appVersions[
+  //     appVersionKey
+  //   ];
+  // }
 
   /**
    * Save app version configuration
    */
   private saveAppVersionConfiguration(
-    appId: string,
-    appVersion: string,
+    appId: number,
+    appVersion: number,
     vincentApp: VincentAppState,
     abilityIpfsCids: string[],
     abilityPolicyCids: string[][]
@@ -356,29 +355,25 @@ export class StateManager {
   }
 
   /**
-   * Get or register Vincent app using app versioning approach
-   */
+     * Get or register Vincent app using app versioning approach
+     */
   async getOrRegisterVincentAppVersioned(
     delegateeAddress: string,
-    registerAppFunction: () => Promise<{ appId: string; appVersion: string }>,
+    registerAppFunction: () => Promise<{ appId: number; appVersion: number }>,
     registerNextVersionFunction: (
-      appId: string
-    ) => Promise<{ appId: string; appVersion: string }>,
+      appId: number
+    ) => Promise<{ appId: number; appVersion: number }>,
     abilityIpfsCids: string[],
     abilityPolicies: string[][],
-    abilityPolicyParameterNames: string[][],
-    abilityPolicyParameterTypes: number[][],
-    abilityPolicyParameterValues?: string[][]
+    policyParams: PermissionData,
   ): Promise<{
-    appId: string;
-    appVersion: string;
+    appId: number;
+    appVersion: number;
     isNew: boolean;
     isNewVersion: boolean;
     abilityIpfsCids: string[];
     abilityPolicies: string[][];
-    abilityPolicyParameterNames: string[][];
-    abilityPolicyParameterTypes: number[][];
-    abilityPolicyParameterValues?: string[][];
+    policyParams: PermissionData;
   }> {
     // First, check if we have any existing app for this delegatee
     const existingApp = this.getExistingVincentApp(delegateeAddress);
@@ -406,9 +401,7 @@ export class StateManager {
           isNewVersion: false,
           abilityIpfsCids,
           abilityPolicies,
-          abilityPolicyParameterNames,
-          abilityPolicyParameterTypes,
-          abilityPolicyParameterValues,
+          policyParams,
         };
       } else {
         // Register new version for existing app
@@ -424,9 +417,7 @@ export class StateManager {
           delegateeAddress,
           abilityIpfsCids,
           abilityPolicies,
-          abilityPolicyParameterNames,
-          abilityPolicyParameterTypes,
-          abilityPolicyParameterValues,
+          policyParams,
           createdAt: new Date().toISOString(),
           network: this.network,
         };
@@ -446,9 +437,7 @@ export class StateManager {
           isNewVersion: true,
           abilityIpfsCids,
           abilityPolicies,
-          abilityPolicyParameterNames,
-          abilityPolicyParameterTypes,
-          abilityPolicyParameterValues,
+          policyParams,
         };
       }
     } else {
@@ -463,9 +452,7 @@ export class StateManager {
         delegateeAddress,
         abilityIpfsCids,
         abilityPolicies,
-        abilityPolicyParameterNames,
-        abilityPolicyParameterTypes,
-        abilityPolicyParameterValues,
+        policyParams,
         createdAt: new Date().toISOString(),
         network: this.network,
       };
@@ -485,9 +472,7 @@ export class StateManager {
         delegateeAddress,
         abilityIpfsCids,
         abilityPolicies,
-        abilityPolicyParameterNames,
-        abilityPolicyParameterTypes,
-        abilityPolicyParameterValues
+        policyParams,
       );
 
       return {
@@ -497,9 +482,7 @@ export class StateManager {
         isNewVersion: false,
         abilityIpfsCids,
         abilityPolicies,
-        abilityPolicyParameterNames,
-        abilityPolicyParameterTypes,
-        abilityPolicyParameterValues,
+        policyParams,
       };
     }
   }
@@ -508,10 +491,10 @@ export class StateManager {
    * Find matching app version based on ability/policy CIDs
    */
   private findMatchingAppVersion(
-    appId: string,
+    appId: number,
     abilityIpfsCids: string[],
     abilityPolicies: string[][]
-  ): string | undefined {
+  ): number | undefined {
     if (!this.nestedState.testFiles[this.testFileName]?.appVersions) {
       return undefined;
     }
@@ -599,12 +582,10 @@ export class StateManager {
   }
 
   generateAccount(): AccountState {
-    const privateKey = generatePrivateKey();
-    const address = privateKeyToAccount(privateKey).address;
-
+    const wallet = ethers.Wallet.createRandom();
     return {
-      privateKey,
-      address,
+      privateKey: wallet.privateKey,
+      address: wallet.address,
       createdAt: new Date().toISOString(),
       network: this.network,
     };
@@ -616,12 +597,10 @@ export class StateManager {
   ): { privateKey: string; address: string; isNew: boolean } {
     // If private key provided from env, use it
     if (existingPrivateKey) {
-      const address = privateKeyToAccount(
-        existingPrivateKey as `0x${string}`
-      ).address;
+      const wallet = new ethers.Wallet(existingPrivateKey);
       return {
         privateKey: existingPrivateKey,
-        address,
+        address: wallet.address,
         isNew: false,
       };
     }
@@ -687,23 +666,23 @@ export class StateManager {
     };
   }
 
-  getGeneratedAccounts(): string[] {
-    // Get accounts for this test file
-    const testFileAccounts =
-      this.nestedState.testFiles[this.testFileName]?.accounts?.[this.network] ||
-      {};
+  // getGeneratedAccounts(): string[] {
+  //   // Get accounts for this test file
+  //   const testFileAccounts =
+  //     this.nestedState.testFiles[this.testFileName]?.accounts?.[this.network] ||
+  //     {};
 
-    // Also check legacy shared accounts for backward compatibility
-    const legacyAccounts =
-      this.nestedState.sharedAccounts?.[this.network] || {};
+  //   // Also check legacy shared accounts for backward compatibility
+  //   const legacyAccounts =
+  //     this.nestedState.sharedAccounts?.[this.network] || {};
 
-    // Combine both sources
-    const allAccounts = { ...legacyAccounts, ...testFileAccounts };
+  //   // Combine both sources
+  //   const allAccounts = { ...legacyAccounts, ...testFileAccounts };
 
-    return Object.entries(allAccounts)
-      .filter(([_, account]) => account !== undefined)
-      .map(([type, account]) => `${type}: ${account!.address}`);
-  }
+  //   return Object.entries(allAccounts)
+  //     .filter(([_, account]) => account !== undefined)
+  //     .map(([type, account]) => `${type}: ${account!.address}`);
+  // }
 
   /**
    * Get existing PKP or return undefined if needs minting
@@ -1015,14 +994,12 @@ export class StateManager {
    * Save a newly registered Vincent app to state
    */
   saveVincentApp(
-    appId: string,
-    appVersion: string,
+    appId: number,
+    appVersion: number,
     delegateeAddress: string,
     abilityIpfsCids: string[],
     abilityPolicies: string[][],
-    abilityPolicyParameterNames: string[][],
-    abilityPolicyParameterTypes: number[][],
-    abilityPolicyParameterValues?: string[][]
+    policyParams: PermissionData,
   ): void {
     const configState = this.getCurrentConfigState();
     configState.state.vincentApp = {
@@ -1031,9 +1008,7 @@ export class StateManager {
       delegateeAddress,
       abilityIpfsCids,
       abilityPolicies,
-      abilityPolicyParameterNames,
-      abilityPolicyParameterTypes,
-      abilityPolicyParameterValues,
+      policyParams,
       createdAt: new Date().toISOString(),
       network: this.network,
     };
@@ -1047,103 +1022,102 @@ export class StateManager {
     console.log(chalk.green(`   ↳ Delegatee: ${delegateeAddress}`));
   }
 
-  /**
-   * Get or register Vincent app for testing
-   */
-  async getOrRegisterVincentApp(
-    delegateeAddress: string,
-    registerFunction: () => Promise<{ appId: string; appVersion: string }>,
-    abilityIpfsCids: string[],
-    abilityPolicies: string[][],
-    abilityPolicyParameterNames: string[][],
-    abilityPolicyParameterTypes: number[][],
-    abilityPolicyParameterValues?: string[][]
-  ): Promise<{
-    appId: string;
-    appVersion: string;
-    isNew: boolean;
-    abilityIpfsCids: string[];
-    abilityPolicies: string[][];
-    abilityPolicyParameterNames: string[][];
-    abilityPolicyParameterTypes: number[][];
-    abilityPolicyParameterValues?: string[][];
-  }> {
-    // Check if we have an existing Vincent app for this delegatee
-    const existingApp = this.getExistingVincentApp(delegateeAddress);
-    if (existingApp) {
-      // Update the current configuration with the new ability/policy information
-      // since we're reusing an existing app but with potentially different abilities
-      const configState = this.getCurrentConfigState();
-      if (configState.state.vincentApp) {
-        console.log(
-          chalk.blue(
-            `🔄 Updating reused app ${existingApp.appId} with current ability/policy configuration`
-          )
-        );
-        configState.state.vincentApp.abilityIpfsCids = abilityIpfsCids;
-        configState.state.vincentApp.abilityPolicies = abilityPolicies;
-        configState.state.vincentApp.abilityPolicyParameterNames =
-          abilityPolicyParameterNames;
-        configState.state.vincentApp.abilityPolicyParameterTypes =
-          abilityPolicyParameterTypes;
-        configState.state.vincentApp.abilityPolicyParameterValues =
-          abilityPolicyParameterValues;
-      }
+  // /**
+  //  * Get or register Vincent app for testing
+  //  */
+  // async getOrRegisterVincentApp(
+  //   delegateeAddress: string,
+  //   registerFunction: () => Promise<{ appId: string; appVersion: string }>,
+  //   abilityIpfsCids: string[],
+  //   abilityPolicies: string[][],
+  //   abilityPolicyParameterNames: string[][],
+  //   abilityPolicyParameterTypes: number[][],
+  //   abilityPolicyParameterValues?: string[][]
+  // ): Promise<{
+  //   appId: string;
+  //   appVersion: string;
+  //   isNew: boolean;
+  //   abilityIpfsCids: string[];
+  //   abilityPolicies: string[][];
+  //   abilityPolicyParameterNames: string[][];
+  //   abilityPolicyParameterTypes: number[][];
+  //   abilityPolicyParameterValues?: string[][];
+  // }> {
+  //   // Check if we have an existing Vincent app for this delegatee
+  //   const existingApp = this.getExistingVincentApp(delegateeAddress);
+  //   if (existingApp) {
+  //     // Update the current configuration with the new ability/policy information
+  //     // since we're reusing an existing app but with potentially different abilities
+  //     const configState = this.getCurrentConfigState();
+  //     if (configState.state.vincentApp) {
+  //       console.log(
+  //         chalk.blue(
+  //           `🔄 Updating reused app ${existingApp.appId} with current ability/policy configuration`
+  //         )
+  //       );
+  //       configState.state.vincentApp.abilityIpfsCids = abilityIpfsCids;
+  //       configState.state.vincentApp.abilityPolicies = abilityPolicies;
+  //       configState.state.vincentApp.abilityPolicyParameterNames =
+  //         abilityPolicyParameterNames;
+  //       configState.state.vincentApp.abilityPolicyParameterTypes =
+  //         abilityPolicyParameterTypes;
+  //       configState.state.vincentApp.abilityPolicyParameterValues =
+  //         abilityPolicyParameterValues;
+  //     }
 
-      return {
-        appId: existingApp.appId,
-        appVersion: existingApp.appVersion,
-        isNew: false,
-        abilityIpfsCids,
-        abilityPolicies,
-        abilityPolicyParameterNames,
-        abilityPolicyParameterTypes,
-        abilityPolicyParameterValues,
-      };
-    }
+  //     return {
+  //       appId: existingApp.appId,
+  //       appVersion: existingApp.appVersion,
+  //       isNew: false,
+  //       abilityIpfsCids,
+  //       abilityPolicies,
+  //       abilityPolicyParameterNames,
+  //       abilityPolicyParameterTypes,
+  //       abilityPolicyParameterValues,
+  //     };
+  //   }
 
-    // Register new Vincent app
-    console.log(chalk.yellow(`🏢 Registering new Vincent app...`));
-    const { appId, appVersion } = await registerFunction();
+  //   // Register new Vincent app
+  //   console.log(chalk.yellow(`🏢 Registering new Vincent app...`));
+  //   const { appId, appVersion } = await registerFunction();
 
-    // Save to state
-    this.saveVincentApp(
-      appId,
-      appVersion,
-      delegateeAddress,
-      abilityIpfsCids,
-      abilityPolicies,
-      abilityPolicyParameterNames,
-      abilityPolicyParameterTypes,
-      abilityPolicyParameterValues
-    );
+  //   // Save to state
+  //   this.saveVincentApp(
+  //     appId,
+  //     appVersion,
+  //     delegateeAddress,
+  //     abilityIpfsCids,
+  //     abilityPolicies,
+  //     abilityPolicyParameterNames,
+  //     abilityPolicyParameterTypes,
+  //     abilityPolicyParameterValues
+  //   );
 
-    return {
-      appId,
-      appVersion,
-      isNew: true,
-      abilityIpfsCids,
-      abilityPolicies,
-      abilityPolicyParameterNames,
-      abilityPolicyParameterTypes,
-      abilityPolicyParameterValues,
-    };
-  }
+  //   return {
+  //     appId,
+  //     appVersion,
+  //     isNew: true,
+  //     abilityIpfsCids,
+  //     abilityPolicies,
+  //     abilityPolicyParameterNames,
+  //     abilityPolicyParameterTypes,
+  //     abilityPolicyParameterValues,
+  //   };
+  // }
 
   /**
    * Update existing Vincent app state with parameter values
    */
   updateVincentAppParameterValues(
-    appId: string,
-    abilityPolicyParameterValues: string[][]
+    appId: number,
+    policyParams: PermissionData,
   ): void {
     const configState = this.getCurrentConfigState();
     if (
       configState.state.vincentApp &&
       configState.state.vincentApp.appId === appId
     ) {
-      configState.state.vincentApp.abilityPolicyParameterValues =
-        abilityPolicyParameterValues;
+      configState.state.vincentApp.policyParams = policyParams;
       console.log(
         chalk.green(
           `💾 Updated Vincent app parameter values for config ${this.configHash}: ${appId}`
@@ -1156,9 +1130,9 @@ export class StateManager {
    * Check if PKP is already permitted for the given app version
    */
   isPKPPermittedForAppVersion(
-    pkpTokenId: string,
-    appId: string,
-    appVersion: string
+    pkpEthAddress: string,
+    appId: number,
+    appVersion: number
   ): boolean {
     const configState = this.getCurrentConfigState();
     if (!configState.state.pkpAppPermissions) {
@@ -1167,7 +1141,7 @@ export class StateManager {
 
     const existingPermission = configState.state.pkpAppPermissions.find(
       (permission) =>
-        permission.pkpTokenId === pkpTokenId &&
+        permission.pkpEthAddress === pkpEthAddress &&
         permission.appId === appId &&
         permission.appVersion === appVersion &&
         permission.network === this.network
@@ -1176,7 +1150,7 @@ export class StateManager {
     if (existingPermission) {
       console.log(
         chalk.blue(
-          `🔐 PKP ${pkpTokenId} already permitted for app ${appId} v${appVersion} in config ${this.configHash}`
+          `🔐 PKP ${pkpEthAddress} already permitted for app ${appId} v${appVersion} in config ${this.configHash}`
         )
       );
       console.log(
@@ -1192,9 +1166,9 @@ export class StateManager {
    * Save PKP app permission to state
    */
   savePKPAppPermission(
-    pkpTokenId: string,
-    appId: string,
-    appVersion: string
+    pkpEthAddress: string,
+    appId: number,
+    appVersion: number
   ): void {
     const configState = this.getCurrentConfigState();
     if (!configState.state.pkpAppPermissions) {
@@ -1204,14 +1178,14 @@ export class StateManager {
     // Check if permission already exists to avoid duplicates
     const existingIndex = configState.state.pkpAppPermissions.findIndex(
       (permission) =>
-        permission.pkpTokenId === pkpTokenId &&
+        permission.pkpEthAddress === pkpEthAddress &&
         permission.appId === appId &&
         permission.appVersion === appVersion &&
         permission.network === this.network
     );
 
     const newPermission: PKPAppPermissionState = {
-      pkpTokenId,
+      pkpEthAddress,
       appId,
       appVersion,
       permittedAt: new Date().toISOString(),
@@ -1223,7 +1197,7 @@ export class StateManager {
       configState.state.pkpAppPermissions[existingIndex] = newPermission;
       console.log(
         chalk.gray(
-          `💾 Updated PKP app permission for config ${this.configHash}: PKP ${pkpTokenId} for app ${appId} v${appVersion}`
+          `💾 Updated PKP app permission for config ${this.configHash}: PKP ${pkpEthAddress} for app ${appId} v${appVersion}`
         )
       );
     } else {
@@ -1231,7 +1205,7 @@ export class StateManager {
       configState.state.pkpAppPermissions.push(newPermission);
       console.log(
         chalk.gray(
-          `💾 Saved PKP app permission for config ${this.configHash}: PKP ${pkpTokenId} for app ${appId} v${appVersion}`
+          `💾 Saved PKP app permission for config ${this.configHash}: PKP ${pkpEthAddress} for app ${appId} v${appVersion}`
         )
       );
     }
